Fix an issue where thr wrong index is being popped
Since the readDir function will add all the elements in lexical order
there is a possibility that the info file might not at index zero. This
commit fixes the issue by searching for the file, which is either in
thhe first or last location. This may pose a probelm when a profile is
named something that lexographically falls in between the sets. Until
then, this issue can be marked as resolved